basic route tests passing
